requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic


{
  "answer": "I would build an event-driven scalable architecture."
}

{
  "answers": [
    "I would build microservices with autoscaling.",
    "I would design modular services with load balancing.",
    "I would optimize bottlenecks using caching."
  ]
}

uvicorn app:app --reload